我是沙雕罗振韩

+++ reckoning/src/main/java/cn/yugutou/reckoning/dto/resp/QueryUserAndTotalResp.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Data
 public class QueryUserAndTotalResp implements Serializable {
     private List  userInfoByNamePhone;
-    private Integer totalNum;
+    private Long totalNum;
 }
